refactor(bet): 최종 수익 반환 메서드의 반환값 변경 및 검증 추가& 테스트 코드 추가

+++ src/test/java/domain/BetTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.model.Player;
 import domain.model.PlayerStatus;
-import domain.service.BlackJackService;
 import org.junit.jupiter.api.Test;
 
 public class BetTest {
@@ -22,56 +21,56 @@ public class BetTest {
     }
 
     @Test
-    void 플레이어_승리_최종수익(){
+    void 플레이어_승리_최종수익() {
         // given
         Player phobi = Player.of("phobi");
         phobi.bet(10000);
         phobi.changeStatus(PlayerStatus.WIN);
 
         // when
-        int finalMoney = (int) phobi.getFinalMoney();
+        int finalMoney = phobi.getFinalMoney();
 
         // then
         assertThat(finalMoney).isEqualTo(10000);
     }
 
     @Test
-    void 플레이어_패배_최종수익(){
+    void 플레이어_패배_최종수익() {
         // given
         Player phobi = Player.of("phobi");
         phobi.bet(10000);
         phobi.changeStatus(PlayerStatus.LOSS);
 
         // when
-        int finalMoney = (int) phobi.getFinalMoney();
+        int finalMoney = phobi.getFinalMoney();
 
         // then
         assertThat(finalMoney).isEqualTo(-10000);
     }
 
     @Test
-    void 플레이어_블랙잭_최종수익(){
+    void 플레이어_블랙잭_최종수익() {
         // given
         Player phobi = Player.of("phobi");
         phobi.bet(10000);
         phobi.changeStatus(PlayerStatus.BLACK_JACK);
 
         // when
-        int finalMoney = (int) phobi.getFinalMoney();
+        int finalMoney = phobi.getFinalMoney();
 
         // then
         assertThat(finalMoney).isEqualTo(15000);
     }
 
     @Test
-    void 플레이어_무승부_최종수익(){
+    void 플레이어_무승부_최종수익() {
         // given
         Player phobi = Player.of("phobi");
         phobi.bet(10000);
         phobi.changeStatus(PlayerStatus.DRAW);
 
         // when
-        int finalMoney = (int) phobi.getFinalMoney();
+        int finalMoney = phobi.getFinalMoney();
 
         // then
         assertThat(finalMoney).isEqualTo(0);
